refactor(chat): conversation-id storage on lib localStorage adapter + hydration extracted to useChatHistoryHydration [preview:none]
